+++ code/game/g_cmds.c
@@ -1366,7 +1366,7 @@ static qboolean G_SayTo( gentity_t *ent, gentity_t *other, int mode, int color, 
 	}
 
 	if(	channel && other->client->chatchannels &&
-		!G_Q3F_ArrayFind( other->client->chatchannels, (int) channel )
+		!G_Q3F_ArrayFind( other->client->chatchannels, (intptr_t) channel )
 		&& ent != other )
 			return( qfalse );	// This is a channel message, but the client isn't on this channel
 
@@ -2179,7 +2179,7 @@ static void G_Q3F_ChannelCommand( gentity_t *ent )
 	// Allow the user to specify which channels they want for team messages.
 
 	char chanbuff[64];
-	int numargs, curr, index;
+	intptr_t numargs, curr, index;
 	q3f_array_t *array;
 	q3f_data_t *data;
 	char *str;
@@ -2201,14 +2201,14 @@ static void G_Q3F_ChannelCommand( gentity_t *ent )
 		{
 			trap_Argv( curr, chanbuff, 64 );
 			q3f_cc_preprocess( chanbuff );
-			if( !G_Q3F_ArrayFind( array, (int) G_Q3F_GetString( chanbuff ) ) )
+			if( !G_Q3F_ArrayFind( array, (intptr_t) G_Q3F_GetString( chanbuff ) ) )
 			{
 				if( array->used >= Q3F_CHANNEL_MAX )
 				{
 					trap_SendServerCommand( ent->s.number, va( "print \"You can only have %d channels at a time.\n\"", Q3F_CHANNEL_MAX ) );
 					break;
 				}
-				G_Q3F_ArrayAdd( array, Q3F_TYPE_STRING, 0, (int) chanbuff );
+				G_Q3F_ArrayAdd( array, Q3F_TYPE_STRING, 0, (intptr_t) chanbuff );
 			}
 			curr++;
 		}
@@ -2258,14 +2258,14 @@ static void G_Q3F_ChannelCommand( gentity_t *ent )
 		{
 			trap_Argv( curr, chanbuff, 64 );
 			q3f_cc_preprocess( chanbuff );
-			if( !G_Q3F_ArrayFind( array, (int) G_Q3F_GetString( chanbuff ) ) )
+			if( !G_Q3F_ArrayFind( array, (intptr_t) G_Q3F_GetString( chanbuff ) ) )
 			{
 				if( array->used >= Q3F_CHANNEL_MAX )
 				{
 					trap_SendServerCommand( ent->s.number, va( "print \"You can only have %d channels at a time.\n\"", Q3F_CHANNEL_MAX ) );
 					break;
 				}
-				G_Q3F_ArrayAdd( array, Q3F_TYPE_STRING, 0, (int) chanbuff );
+				G_Q3F_ArrayAdd( array, Q3F_TYPE_STRING, 0, (intptr_t) chanbuff );
 			}
 			curr++;
 		}
@@ -3561,8 +3561,8 @@ static void G_Q3F_PlayerStatus( gentity_t *ent )
 		}
 		else if( player->client && player->client->pers.connected == CON_CONNECTING )
 		{
-			trap_SendServerCommand( ent-g_entities, va("print \"%3d %4d %6s %5d %s\n\"",
-								(int)(player-g_entities),
+			trap_SendServerCommand( ent-g_entities, va("print \"%3ld %4d %6s %5d %s\n\"",
+								(intptr_t)(player-g_entities),
 								player->client->ps.ping,
 								"conn",
 								player->client->ps.persistant[PERS_SCORE],
